Extract carousel initializer to a module
